task-03-RAG - Part 1 (Basic Document QA)/requirements.txt
pyyaml
pypdf
langchain
langchain_community
faiss-cpu
sentence-transformers
transformers
tqdm
streamlit
python-dotenv
google-generativeai
langchain_core